safari-534.53-branch/LayoutTests/media/video-muted-expected.txt

Test 'muted' attribute

EXPECTED (video.muted == 'false') OK
RUN(video.muted = true)
EXPECTED (video.muted == 'true') OK
RUN(video.load())
EVENT(canplaythrough)
EXPECTED (video.muted == 'true') OK
RUN(video.muted = false)
EXPECTED (video.muted == 'false') OK
END OF TEST
